More robust git script

+++ systemd-sync-master.sh
@@ -5,8 +5,6 @@
 # sudo systemctl start  dot-files-master-sync.service
 # sudo systemctl stop   dot-files-master-sync.service
 
-set -e
-
 if [ ! "$(uname)" == "Linux" ]; then echo "Must be on a Linux system."; exit 1; fi
 
 script_dir="$(dirname "$(readlink -f "$0")")"
@@ -14,11 +12,16 @@ curr_dir=$PWD
 
 while true
 do
-    cd "$script_dir" || exit 1
-    git fetch
-    git reset --hard origin/master
-    cd "$curr_dir" || exit 1
-    echo SLEEPING FOR 60s
-    date
+    cd "$script_dir" || cd "$curr_dir" || exit 1
+
+    if wget -q --spider http://google.com; then
+        printf "Online: "
+        git fetch
+        git reset --hard origin/master
+    else
+        printf "Offline: "
+    fi
+
+    printf "Sleeping for 60s at: %s\n" "$(date +%H:%M:%S)"
     sleep 60
 done
